+++ src/Sharpen.VisualStudioExtension/ToolWindows/AnalysisResultTreeViewItems/CSharpVersionTreeViewItem.cs
@@ -1,4 +1,5 @@
-﻿using Sharpen.VisualStudioExtension.ToolWindows.AnalysisResultTreeViewBuilders;
+﻿using Sharpen.Engine;
+using Sharpen.VisualStudioExtension.ToolWindows.AnalysisResultTreeViewBuilders;
 
 namespace Sharpen.VisualStudioExtension.ToolWindows.AnalysisResultTreeViewItems
 {
@@ -6,10 +7,15 @@ namespace Sharpen.VisualStudioExtension.ToolWindows.AnalysisResultTreeViewItems
     {
         public string CSharpVersion { get; }
 
-        public CSharpVersionTreeViewItem(BaseTreeViewItem parent, IAnalysisResultTreeViewBuilder treeViewBuilder, int numberOfItems, string csharpVersion)
-            : base(parent, treeViewBuilder, numberOfItems, "C# " + csharpVersion, null) // TODO: Replace the string creation once the abstraction for the C# version is fully implemented.
+        public CSharpVersionTreeViewItem(BaseTreeViewItem parent, IAnalysisResultTreeViewBuilder treeViewBuilder, int numberOfItems, string cSharpVersion)
+            : base(parent, treeViewBuilder, numberOfItems, "C# " + cSharpVersion, CSharpLanguageVersions.GetWhatIsNewUrlFor(cSharpVersion), GetLearnMoreDisplayTextFor(cSharpVersion)) // TODO: Replace the string creation once the abstraction for the C# version is fully implemented.
         {
-            CSharpVersion = csharpVersion;
+            CSharpVersion = cSharpVersion;
+        }
+
+        private static string GetLearnMoreDisplayTextFor(string cSharpVersion)
+        {
+            return $"What {(cSharpVersion == CSharpLanguageVersions.Latest ? "is" : "was")} new in C# {cSharpVersion}?";
         }
     }
 }